+++ packages/azure/CHANGELOG.md
@@ -1,5 +1,43 @@
 # @ai-sdk/azure
 
+## 3.0.7
+
+### Patch Changes
+
+- Updated dependencies [89202fb]
+  - @ai-sdk/openai@3.0.7
+
+## 3.0.6
+
+### Patch Changes
+
+- Updated dependencies [dc87517]
+  - @ai-sdk/openai@3.0.6
+
+## 3.0.5
+
+### Patch Changes
+
+- Updated dependencies [d937c8f]
+  - @ai-sdk/provider@3.0.2
+  - @ai-sdk/openai@3.0.5
+  - @ai-sdk/provider-utils@4.0.4
+
+## 3.0.4
+
+### Patch Changes
+
+- Updated dependencies [0b429d4]
+  - @ai-sdk/provider-utils@4.0.3
+  - @ai-sdk/openai@3.0.4
+
+## 3.0.3
+
+### Patch Changes
+
+- Updated dependencies [55cd1a4]
+  - @ai-sdk/openai@3.0.3
+
 ## 3.0.2
 
 ### Patch Changes